+++ src/main/java/ru/skypro/courseWork/service/impl/CommentServiceImpl.java
@@ -15,6 +15,7 @@ import ru.skypro.courseWork.mapper.CommentMapper;
 import ru.skypro.courseWork.repository.AdRepository;
 import ru.skypro.courseWork.repository.CommentRepository;
 import ru.skypro.courseWork.repository.UserRepository;
+import ru.skypro.courseWork.security.service.SecurityUtils;
 import ru.skypro.courseWork.service.CommentService;
 
 import javax.transaction.Transactional;
@@ -30,6 +31,7 @@ public class CommentServiceImpl implements CommentService {
     private final CommentMapper commentMapper;
     private final AdRepository adRepository;
     private final UserRepository userRepository;
+    private final SecurityUtils securityUtils;
     @Override
     public List<CommentDto> getCommentByIdAd(Integer id) {
         return commentMapper.toCommentsDto(commentRepository.findAllByAdPk(id));
@@ -42,7 +44,6 @@ public class CommentServiceImpl implements CommentService {
         Ad ad = adRepository.findById(id).orElseThrow(AdNotFoundException::new);
         Comment comment = commentMapper.toCommentEntityFromCreateOrUpdateComment(createOrUpdateCommentDto);
 
-        //comment.setCreateAt(Instant.now().toEpochMilli());
         comment.setAd(ad);
         comment.setAuthor(user);
 
@@ -52,6 +53,10 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public void deleteCommentById(Integer commentId) {
+
+        Comment comment = commentRepository.findById(commentId).orElseThrow(CommentNotFoundException::new);
+
+        securityUtils.checkAccessToComment(comment);
         commentRepository.deleteById(commentId);
     }
 
@@ -64,6 +69,8 @@ public class CommentServiceImpl implements CommentService {
         User user = userRepository.findByEmail(authentication.getName()).orElseThrow(UserNotFoundException::new);
         Ad ad = adRepository.findById(adId).orElseThrow(AdNotFoundException::new);
         Comment comment = commentRepository.findById(commentId).orElseThrow(CommentNotFoundException::new);
+
+        securityUtils.checkAccessToComment(comment);
 
         comment.setText(createOrUpdateCommentDto.getText());
         comment.setCreateAt(Instant.now().toEpochMilli());
